gateway/onvif-discovery/requirements.txt
onvif-zeep>=0.2.12
WSDiscovery>=2.0.0
requests>=2.31
pyyaml>=6.0
